+++ packages/stream_chat_flutter/lib/stream_chat_flutter.dart
@@ -157,5 +157,6 @@ export 'src/utils/device_segmentation.dart';
 export 'src/utils/extensions.dart';
 export 'src/utils/helpers.dart';
 export 'src/utils/message_preview_formatter.dart';
+export 'src/utils/network_error_text.dart';
 export 'src/utils/stream_image_cdn.dart';
 export 'src/utils/typedefs.dart';